requirements.txt
flask
python-dotenv
pymongo
bcrypt
gunicorn